add searchtype option for simpleAL
searchtype `grid` is the current default behaviour of creating a
searchspace using a meshgrid over all input variables ->
`nsearch^ndim` points. This gets too expensive with a moderate number
of input dimensions.
searchtype `halton` is a new option which distributes a fixed number
`nsearch` of samples pseudo-randomly within the searchspace